requirements.txt
asgiref>=3.8,<4.0
boto3>=1.35,<1.36
dj-database-url>=2.2,<2.3
Django>=5.1,<5.2
django-cors-headers>=4.4,<4.5
django-environ>=0.11,<0.12
django-storages>=1.14,<1.15
djangorestframework>=3.15,<3.16
gunicorn>=22,<24
Pillow>=10.4,<10.5
psycopg2-binary>=2.9.9,<3.0
sqlparse>=0.5,<0.6
